# yassine_olive_mill_backend-fedi/yassine_olive_mill_backend-fedi/api_documentation.yaml
openapi: 3.0.0
info:
  title: Olive Oil Mill API
  version: 1.0.0
  description: API for olive oil mill management system
servers:
  - url: http://localhost:3000/api
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /invoices:
    get:
      security:
        - BearerAuth: []
      summary: Get all invoices
      description: Requires admin or manager role
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
        - name: status
          in: query
          schema: { type: string, enum: [draft, sent, paid, overdue] }
        - name: clientId
          in: query
          schema: { type: integer }
      responses:
        200:
          description: Paginated invoices with client, batch, payments
    post:
      summary: Create invoice
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [clientId, amount]
              properties:
                clientId: { type: integer }
                batchId: { type: integer }
                processing_decisionId: { type: integer }
                amount: { type: number }
                due_date: { type: string, format: date }
                notes: { type: string }
      responses:
        201:
          description: Invoice created with status 'draft'
  /invoices/{id}/status:
    put:
      summary: Update invoice status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [draft, sent, paid, overdue] }
      responses:
        200:
          description: Invoice status updated
        404:
          description: Invoice not found
  /payments:
    get:
      summary: Get all payments
      parameters:
        - name: invoiceId
          in: query
          schema: { type: integer }
      responses:
        200:
          description: Payments with invoice info
    post:
      summary: Record payment
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [invoiceId, amount, payment_method]
              properties:
                invoiceId: { type: integer }
                amount: { type: number }
                payment_method: { type: string }
                reference: { type: string }
      responses:
        201:
          description: Payment recorded, invoice status updated if fully paid
        404:
          description: Invoice not found
...